gateway/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.1
pydantic-settings==2.1.0
email-validator==2.1.0  # Add this explicitly
starlette==0.27.0
python-jose==3.3.0
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.1
redis==5.0.1
prometheus-client==0.19.0
cryptography==41.0.5
bcrypt==4.0.1
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.0